Brought to completion - removed debug things - moved to more appropriate spot in guis/panel/hv - enhanced id generation
